requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy>=2.0,<3.0
jinja2==3.1.6
google-genai>=1.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
